requirements.txt
pydantic
email-validator
alembic
amqp
annotated-types
anyio
bcrypt
billiard
celery
click
click-didyoumean
click-plugins
click-repl
diskcache
dnspython
python-docx
fastapi
greenlet
h11
idna
iniconfig
Jinja2
kombu
llama-cpp-python
lxml
Mako
MarkupSafe
numpy
packaging
passlib[bcrypt]
pillow
pluggy
prompt_toolkit
psycopg2-binary
pydantic_core
pydantic-settings
PyJWT
PyPDF2
pytest
python-dateutil
python-dotenv
python-multipart
redis
six
sniffio
SQLAlchemy
starlette
typing_extensions
tzdata
uvicorn
vine
wcwidth
httpx